feat: OMR currency support with 3 decimals and schema updates

+++ src/app/api/admin/pos/shift/close/route.ts
@@ -10,7 +10,7 @@ const closeShiftSchema = z.object({
   shiftId: z.string().min(1, "Shift ID is required"),
   denominations: z.array(
     z.object({
-      value: z.number().int().positive("Denomination value must be positive"),
+      value: z.number().positive("Denomination value must be positive"),
       quantity: z.number().int().nonnegative("Quantity must be a non-negative number"),
     })
   ).min(1, "At least one denomination entry is required"),
